backend/requirements.txt
pymongo[srv]
python-dotenv
bcrypt
sentence-transformers
chromadb
python-dateutil
groq
